ESRF data policy - metadata upload: - Moved code from ESRFMultiCollect to ESRFMetadataManager - Call of ESRFMultiCollect.data_collection_hook from ID30BMultiCollect
