+++ ESRF/ESRFMetadataManagerClient.py
@@ -3,8 +3,13 @@
 """A simple client for MetadataManager and MetaExperiment
 """
 
+import os
 import sys
+import logging
 import PyTango.client
+import traceback
+from email.mime.text import MIMEText
+import smtplib
 
 class MetadataManagerClient(object):
 
@@ -121,14 +126,237 @@ class MetadataManagerClient(object):
             print "Unexpected error:", sys.exc_info()[0]
             raise        
 
+    def getState(self):
+        return str(MetadataManagerClient.metadataManager.state())
+
+class MXCuBEMetadataClient(object):
+    
+    def __init__(self, esrf_multi_collect):
+        self.esrf_multi_collect = esrf_multi_collect
+        self._metadataManagerName = self.esrf_multi_collect.getProperty("metadata_manager_name")
+        self._metaExperimentName  = self.esrf_multi_collect.getProperty("meta_experiment_name")
+        self._sessionObject = self.esrf_multi_collect.getObjectByRole("session")
+
+    
+    def reportStackTrace(self):
+        (exc_type, exc_value, exc_traceback) = sys.exc_info()
+        errorMessage = "{0} {1}".format(exc_type, exc_value)
+        errorMessage += "\n\n"
+        listTrace = traceback.extract_tb(exc_traceback)
+        errorMessage += "Traceback (most recent call last): \n"
+        for listLine in listTrace:
+            errorMessage += "  File \"%s\", line %d, in %s%s\n" % (listLine[0], listLine[1], listLine[2], os.linesep)
+        COMMASPACE = ', '
+        mime_text_message = MIMEText(errorMessage)
+        replyTo = "svensson@esrf.fr"
+        listTo = ["svensson@esrf.fr"]
+        listCC = []
+        listBCC = []
+        mime_text_message['Subject'] = "Metadata upload error on MASSIF 1"
+        mime_text_message['From'] = replyTo
+        mime_text_message['To'] = COMMASPACE.join(listTo)
+        if len(listCC) > 0:
+            mime_text_message['CC'] = COMMASPACE.join(listCC)
+        if len(listBCC) > 0:
+            mime_text_message['BCC'] = COMMASPACE.join(listBCC)
+        try:
+            smtp = smtplib.SMTP("localhost")
+            smtp.sendmail(replyTo, listTo + listCC + listBCC, mime_text_message.as_string())
+            smtp.quit()
+        except:
+            pass
+        return errorMessage
+        
+
+    def start(self, data_collect_parameters):
+        # Metadata
+        if self._metadataManagerName is not None and self._metaExperimentName is not None:
+            try:
+                # Create proxy object
+                self._metadataManagerClient = MetadataManagerClient(self._metadataManagerName, self._metaExperimentName)
+                
+                # First check the state of the device server
+                serverState = self._metadataManagerClient.getState()
+                if serverState == "RUNNING":
+                    # Force end of scan
+                    self._metadataManagerClient.end()
+                    
+                fileinfo =  data_collect_parameters["fileinfo"]
+                directory = fileinfo["directory"]
+                prefix = fileinfo["prefix"]
+                run_number = int(fileinfo["run_number"])
+                # Connect to ICAT metadata database
+                
+                proposal = self._sessionObject.get_proposal()
+                
+                # Strip the prefix from any workflow expTypePrefix
+                # TODO: use the ISPyB sample name instead
+                sampleName = prefix
+                for expTypePrefix in ["line-", "mesh-", "ref-", "burn-", "ref-kappa-"]:
+                    if sampleName.startswith(expTypePrefix):
+                        sampleName = sampleName.replace(expTypePrefix, "")
+                        break
+                # The data set name must be unique so we use the ISPyB data collection id
+                datasetName = "{0}_{1}_{2}".format(prefix, run_number, self.esrf_multi_collect.collection_id)
+                self._metadataManagerClient.start(directory, proposal, sampleName, datasetName)
+                self._metadataManagerClient.printStatus()        
+            except:
+                logging.getLogger("user_level_log").warning("Cannot connect to metadata server")
+                errorMessage = self.reportStackTrace()
+                logging.getLogger("user_level_log").warning(errorMessage)
+                self._metadataManagerClient = None
+
+
+    def upload_images_to_icat(self, template, prefix, run_number, directory, 
+                              number_of_images, start_image_number, overlap):
+        logging.getLogger("user_level_log").info("Uploading to images to ICAT")
+        for index in range(number_of_images):
+            image_no = index + start_image_number
+            image_path = os.path.join(directory, template % image_no)
+            self._metadataManagerClient.appendFile(image_path)
+
+ 
+    def end(self, data_collect_parameters):
+        try:
+            dictMetadata = self.getMetadata(data_collect_parameters)
+            if self._metadataManagerClient is not None:
+                # Upload all images
+                fileinfo = data_collect_parameters["fileinfo"]
+                prefix = fileinfo["prefix"]
+                template = fileinfo["template"]
+                directory = fileinfo["directory"]
+                run_number = int(fileinfo["run_number"])
+                for oscillation_parameters in data_collect_parameters["oscillation_sequence"]:
+                    number_of_images = oscillation_parameters["number_of_images"]
+                    start_image_number = oscillation_parameters["start_image_number"]
+                    overlap = oscillation_parameters["overlap"]
+                    self.upload_images_to_icat(template, prefix, run_number, directory, 
+                                               number_of_images, start_image_number, overlap)
+                # Upload the two paths to the meta data HDF5 files
+                beamline = self._sessionObject.endstation_name
+                proposal = self._sessionObject.get_proposal()
+                pathToHdf5File1 = os.path.join(directory, 
+                                               "{proposal}-{beamline}-{prefix}_{run_number}_{dataCollectionId}.h5".format(
+                                                    beamline=beamline,
+                                                    proposal=proposal,
+                                                    prefix=prefix,
+                                                    run_number=run_number,
+                                                    dataCollectionId=self.esrf_multi_collect.collection_id)
+                                               )
+                self._metadataManagerClient.appendFile(pathToHdf5File1)
+                pathToHdf5File2 = os.path.join(directory, 
+                                               "{proposal}-{prefix}-{prefix}_{run_number}_{dataCollectionId}.h5".format(
+                                                    proposal=proposal,
+                                                    prefix=prefix,
+                                                    run_number=run_number,
+                                                    dataCollectionId=self.esrf_multi_collect.collection_id)
+                                               )
+                self._metadataManagerClient.appendFile(pathToHdf5File2)
+                # Upload meta data as attributes
+                # These attributes are common for all ESRF MX beamlines
+                dictMetadata = self.getMetadata(data_collect_parameters)
+                #import pprint
+                #pprint.pprint(dictMetadata)
+                for attributeName, value in dictMetadata.iteritems():
+                    logging.getLogger("HWR").info("Setting metadata client attribute '{0}' to '{1}'".format(attributeName, value))
+                    setattr(self._metadataManagerClient.metadataManager, attributeName, str(value))
+                self._metadataManagerClient.printStatus()
+                self._metadataManagerClient.end()
+        except:
+            logging.getLogger("user_level_log").warning("Cannot upload metadata")
+            errorMessage = self.reportStackTrace()
+            logging.getLogger("user_level_log").warning(errorMessage)
+            self._metadataManagerClient = None
+            
+
+    def getMetadata(self, data_collect_parameters):
+        """
+        Common metadata parameters for ESRF MX beamlines.
+        """
+        listAttributes = [
+                          ["MX_beamShape", "beamShape"], 
+                          ["MX_beamSizeAtSampleX", "beamSizeAtSampleX"], 
+                          ["MX_beamSizeAtSampleY", "beamSizeAtSampleY"], 
+                          ["MX_dataCollectionId", "collection_id"],  
+                          ["MX_directory", "fileinfo.directory"],
+                          ["MX_exposureTime", "oscillation_sequence.exposure_time"],
+                          ["MX_flux", "flux"], 
+                          ["MX_fluxEnd", "flux_end"],
+                          ["MX_numberOfImages", "oscillation_sequence.number_of_images"],
+                          ["MX_oscillationRange", "oscillation_sequence.range"],
+                          ["MX_oscillationStart", "oscillation_sequence.start"],
+                          ["MX_oscillationOverlap", "oscillation_sequence.overlap"],
+                          ["MX_resolution", "resolution"], 
+                          ["MX_startImageNumber", "oscillation_sequence.start_image_number"],
+                          ["MX_scanType", "experiment_type"], 
+                          ["MX_template", "fileinfo.template"],
+                          ["MX_transmission", "transmission"], 
+                          ["MX_xBeam", "xBeam"], 
+                          ["MX_yBeam", "yBeam"], 
+                          ["InstrumentMonochromator_wavelength", "wavelength"],
+                          ]
+        dictMetadata = {}
+        for attribute in listAttributes:
+            if type(attribute) == list:
+                attributeName = attribute[0]
+                keyName = attribute[1]
+            else:
+                attributeName = attribute
+                keyName = attribute
+            value = None
+            if "." in keyName:
+                parent, child = str(keyName).split(".")
+                parentObject = data_collect_parameters[parent]
+                if type(parentObject) == type([]):
+                    parentObject = parentObject[0]
+                if child in parentObject:
+                    value = str(parentObject[child])
+            elif keyName in data_collect_parameters:
+                value = str(data_collect_parameters[keyName])
+            if value is not None:
+                dictMetadata[attributeName] = value
+        # Template - replace python formatting with hashes
+        dictMetadata["MX_template"] = dictMetadata["MX_template"].replace("%04d", "####")
+        # Motor positions
+        motorNames = ""
+        motorPositions = ""
+        for motor, position in data_collect_parameters["motors"].iteritems():
+            if type(motor) == str:
+                motorName = motor
+            else:
+                 nameAttribute = getattr(motor, "name")
+                 if type(nameAttribute) == str:
+                     motorName = nameAttribute
+                 else:
+                     motorName = nameAttribute()
+            if motorNames == "":
+                motorNames = motorName
+                motorPositions = str(round(position, 3))
+            else:
+                motorNames += " " + motorName
+                motorPositions += " " + str(round(position, 3))
+        dictMetadata["MX_motors_name"] = motorNames
+        dictMetadata["MX_motors_value"] = motorPositions
+          # Detector distance
+        distance = self.esrf_multi_collect.get_detector_distance()
+        if distance is not None:
+            dictMetadata["MX_detectorDistance"] = distance
+        # Aperture
+        if self.esrf_multi_collect.bl_control.beam_info is not None and self.esrf_multi_collect.bl_control.beam_info.aperture_hwobj is not None:
+            aperture = self.esrf_multi_collect.bl_control.beam_info.aperture_hwobj.getPosition()
+            dictMetadata["MX_aperture"] =  aperture
+        return dictMetadata
+
+
+
 if __name__ == '__main__':
     metadataManagerName = 'id30a1/metadata/ingest'
     metaExperimentName = 'id30a1/metadata/experiment'
     client = MetadataManagerClient(metadataManagerName, metaExperimentName)
     
-    client.start('/data/visitor/mx415/id30a2/20160909/RAW_DATA', 'mx415', 'sample2', 'dataset_20160909_1')
-    client.appendFile('/data/visitor/mx415/id30a2/20160909/RAW_DATA/t1/test1.txt')
-    client.appendFile('/data/visitor/mx415/id30a2/201600909/RAW_DATA/t1/test2.txt')
+    client.start('/data/visitor/mx415/id30a1/20161014/RAW_DATA', 'mx415', 'sample1', 'dataset_20161014_1')
+    client.appendFile('/data/visitor/mx415/id30a1/20161014/RAW_DATA/t1/test1.txt')
+    client.appendFile('/data/visitor/mx415/id30a1/20161014/RAW_DATA/t1/test2.txt')
     client.printStatus()
     client.end()
